+++ apps/wing/src/analytics/collectors/cli-collector.ts
@@ -1,0 +1,29 @@
+import { Command } from "commander";
+import { Collector } from "./collector";
+
+export interface CLIData {
+  target: string;
+  version: string;
+  options: string;
+}
+
+export class CLICollector extends Collector {
+  private cmd: Command;
+
+  constructor(cmd: Command) {
+    super();
+    this.cmd = cmd;
+  }
+
+  async canCollect(): Promise<boolean> {
+    return true;
+  }
+
+  async collect(): Promise<CLIData> {
+    return {
+      target: this.cmd.opts().target,
+      options: `${JSON.stringify(this.cmd.opts())}`,
+      version: this.cmd.opts().version,
+    }
+  }
+}
