+++ apps/wing/src/analytics/collectors/os-collector.ts
@@ -1,0 +1,22 @@
+import { Collector } from "./collector";
+import os from "os";
+
+export interface OSData {
+  arch: string;
+  platform: string;
+  release: string;
+}
+
+export class OSCollector extends Collector {
+  async canCollect(): Promise<boolean> {
+    return true;
+  }
+
+  async collect(): Promise<OSData> {
+    return {
+      arch: os.arch(),
+      platform: os.platform(),
+      release: os.release(),
+    }
+  }
+}
