+++ apps/wing/src/analytics/export.ts
@@ -1,7 +1,7 @@
 import { spawn } from 'child_process';
 
 export function exportAnalytics(filePath: string) {
-  const child = spawn(process.argv[0], [require.resolve('./exporter'), filePath], {
+  const child = spawn(process.argv[0], [require.resolve('./scripts/detached-export'), filePath], {
     detached: true,
     stdio: 'ignore',
     windowsHide: true,
